requirements.txt
notebook
pandas
scikit-learn
apache-airflow-providers-amazon
mlflow

